Added "nullpointer check" in the draw() method in GamesMenu and MainMenu.

+++ src/com/example/arcade/utilities/Constants.java
@@ -1,6 +1,6 @@
 package com.example.arcade.utilities;
 
 public class Constants {
-    public static int WINDOW_HEIGHT, WINDOW_WIDTH;
+    public static float WINDOW_HEIGHT, WINDOW_WIDTH;
 
 }
